Extract xterm init effect into terminal-surface.ts (#tech-debt)
Move the ~310-line xterm initialization useEffect from use-terminal.ts
into a createTerminalSurface() function in a new terminal-surface.ts
module. This includes XTerm instance creation, addon loading, clipboard
handling, touch/mouse event wiring, keyboard input forwarding, and
resize observation. use-terminal.ts drops from 1158 to 846 lines.
